Revamped the typing system.
Implemented isAssignableFrom almost according to spec.
Refactored clients of the typing system.
Added accurate Constant computation.
Rebuilt a couple of inspections.
Removed the GoUnderlyingType tree.
A lot of other cleanups.
Also Fixes #858

+++ src/ro/redeul/google/go/lang/psi/expressions/literals/GoLiteralFloat.java
@@ -1,5 +1,7 @@
 package ro.redeul.google.go.lang.psi.expressions.literals;
 
-public interface GoLiteralFloat extends GoLiteral<Float> {
+import java.math.BigDecimal;
+
+public interface GoLiteralFloat extends GoLiteral<BigDecimal> {
 
 }
